# docs/swagger.yaml
basePath: /v0
definitions:
  customer.CreateCustomerRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        description: ID          string  `json:"id,omitempty"`
        maxLength: 30
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  customer.CustomerLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  customer.CustomerLoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/customer.CustomerResponse'
    type: object
  customer.CustomerResponse:
    properties:
      email:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
host: api.coffee-culture.uk
info:
  contact: {}
  description: This is the REST API for Coffee Culture.
  title: Coffee Culture API
  version: "1"
paths:
  /customer/create:
    post:
      consumes:
      - application/json
      description: Create a new customer
      parameters:
      - description: Customer object to be created
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/customer.CreateCustomerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Customer created successfully
          schema:
            $ref: '#/definitions/customer.CustomerResponse'
        "400":
          description: Invalid request format or parameters
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new customer
      tags:
      - customers
  /customer/login:
    post:
      consumes:
      - application/json
      description: Login customer
      parameters:
      - description: Customer login credentials
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/customer.CustomerLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Customer logged in successfully
          schema:
            $ref: '#/definitions/customer.CustomerLoginResponse'
        "400":
          description: Invalid request format or parameters
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Login customer
      tags:
      - customers
schemes:
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"